requirements.txt
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.19.0,<2.0.0

